# src/main/resources/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "0.1.0",
    "title": "Conveyancer API",
    "description": "HM Land Registry Digital Street Proof of Concept",
    "contact": {
      "name": "HM Land Registry"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
    "/titles": {
      "get": {
        "summary": "Retrieve a list of Titles",
        "operationId": "get_titles",
        "tags": [
          "Titles"
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TitleTransfer"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/titles/{title_number}": {
      "get": {
        "summary": "Retrieve a specific Title",
        "operationId": "get_title",
        "tags": [
          "Titles"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleTransfer"
                }
              }
            },
            "links": {
              "AddRestriction": {
                "$ref": "#/components/links/AddRestriction"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "summary": "Request a Title to be issued",
        "operationId": "request_title",
        "tags": [
          "Titles"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "requestBody": {
          "description": "Owner to request for",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TitleOwner"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Title issuance has been successfully requested",
            "links": {
              "GetTitle": {
                "$ref": "#/components/links/GetTitle"
              },
              "DraftSalesAgreement": {
                "$ref": "#/components/links/DraftSalesAgreement"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/titles/{title_number}/sales-agreement": {
      "get": {
        "summary": "Retrieve a specific Title's Sales Agreement",
        "operationId": "get_sales_agreement",
        "tags": [
          "Sales Agreement"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesAgreement"
                }
              }
            },
            "links": {
              "SignSalesAgreement": {
                "$ref": "#/components/links/SignSalesAgreement"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "summary": "Update the Title's Sales Agreement",
        "operationId": "update_sales_agreement",
        "tags": [
          "Sales Agreement"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "requestBody": {
          "description": "Title data to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SalesAgreementSign"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesAgreement"
                }
              }
            },
            "links": {
              "GetSalesAgreement": {
                "$ref": "#/components/links/GetSalesAgreement"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "summary": "Drafts a Sales Agreement for a Title",
        "operationId": "add_sales_agreement_to_title",
        "tags": [
          "Sales Agreement"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "requestBody": {
          "description": "New Sales Agreement data to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesAgreementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created Sales Agreement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesAgreement"
                }
              }
            },
            "links": {
              "GetSalesAgreement": {
                "$ref": "#/components/links/GetSalesAgreement"
              },
              "SignSalesAgreement": {
                "$ref": "#/components/links/SignSalesAgreement"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/titles/{title_number}/restrictions": {
      "get": {
        "summary": "Retrieve a list of Restrictions on a Title",
        "operationId": "get_restrictions_on_title",
        "tags": [
          "Restrictions"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          },
          {
            "$ref": "#/components/parameters/RestrictionType"
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Restriction"
                      },
                      {
                        "$ref": "#/components/schemas/ChargeRestriction"
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "summary": "Adds a Restriction to a Title",
        "description": "If the type of Restriction is `CBCR`, a Charge Restriction will be added to the Title in addition.",
        "operationId": "add_restriction_to_title",
        "tags": [
          "Restrictions"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "requestBody": {
          "description": "New Restriction data to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Restriction"
                  },
                  {
                    "$ref": "#/components/schemas/ChargeRestriction"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created Restriction",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Restriction"
                    },
                    {
                      "$ref": "#/components/schemas/ChargeRestriction"
                    }
                  ]
                }
              }
            },
            "links": {
              "RemoveRestriction": {
                "$ref": "#/components/links/RemoveRestrictions"
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/titles/{title_number}/charges": {
      "get": {
        "summary": "Retrieve a list of Charges on a Title",
        "operationId": "get_charges_on_title",
        "tags": [
          "Charges"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Charge"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "summary": "Update the Charges on a Title",
        "operationId": "update_charges",
        "tags": [
          "Charges"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TitleNumber"
          }
        ],
        "requestBody": {
          "description": "What to update on the Charges",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "request_discharge"
                    ],
                    "example": "request_discharge"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request."
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TitleTransfer": {
        "required": [
          "title_number",
          "states",
          "status"
        ],
        "properties": {
          "title_number": {
            "type": "string",
            "pattern": "^([A-Z]{0,3}[1-9][0-9]{0,5}|[0-9]{1,6}[ZT])$",
            "example": "ZQV888860"
          },
          "title": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Title"
              }
            ],
            "nullable": true
          },
          "proposed_title": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Title"
              }
            ],
            "nullable": true
          },
          "sales_agreement": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SalesAgreement"
              }
            ],
            "nullable": true
          },
          "states": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StateSummary"
            },
            "example": {
              "land_title": {
                "state_status": "issued",
                "timestamp": "2018-11-07T10:25:29+00:00"
              },
              "sales_agreement": {
                "state_status": "created",
                "timestamp": "2018-11-08T10:25:29+00:00"
              },
              "request_issuance": {
                "state_status": "approved",
                "timestamp": "2018-11-06T10:25:29+00:00"
              }
            }
          },
          "status": {
            "type": "string",
            "example": "land_title_issued"
          }
        }
      },
      "StateSummary": {
        "required": [
          "state_status",
          "timestamp"
        ],
        "properties": {
          "state_status": {
            "type": "string",
            "example": "approved"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2018-11-07T10:25:29+00:00"
          }
        }
      },
...
      "TitleOwner": {
        "properties": {
          "title_number": {
            "$ref": "#/components/schemas/TitleTransfer/properties/title_number"
          },
          "owner": {
            "$ref": "#/components/schemas/Owner"
          }
        }
      },
...
      "SalesAgreement": {
        "properties": {
          "sdlt": {
            "type": "number",
            "format": "double",
            "example": 1125.01,
            "readOnly": true
          },
          "sdlt_currency_code": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "GBP",
            "readOnly": true
          }
        }
      },
      "SalesAgreementSign": {
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "sign"
            ]
          },
          "signatory": {
            "$ref": "#/components/schemas/X500"
          },
          "signatory_individual": {
            "$ref": "#/components/schemas/Person"
          }
        },
        "additionalProperties": false
      },
...
      "Restriction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RestrictionType"
          },
          {
            "$ref": "#/components/schemas/RestrictionBase"
          }
        ]
      },
      "ChargeRestriction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RestrictionType"
          },
          {
            "$ref": "#/components/schemas/RestrictionBase"
          }
        ],
        "required": [
          "charge"
        ],
        "properties": {
          "charge": {
            "$ref": "#/components/schemas/Charge"
          }
        }
      },
...
      "Owner": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Person"
          }
        ]
      },
...
    },
    "parameters": {
      "TitleNumber": {
        "name": "title_number",
        "in": "path",
        "required": true,
        "description": "The unique ID of a Title",
        "schema": {
          "type": "string",
          "pattern": "^([A-Z]{0,3}[1-9][0-9]{0,5}|[0-9]{1,6}[ZT])$"
        }
      },
      "RestrictionType": {
        "name": "type",
        "in": "query",
        "description": "The restriction_type of a Restriction to filter by",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "ORES",
              "CBCR"
            ]
          }
        }
      }
    },
    "responses": {
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "links": {
      "GetTitle": {
        "operationId": "get_title",
        "parameters": {
          "title_number": "$response.body#/title_number"
        }
      },
      "GetSalesAgreement": {
        "operationId": "get_sales_agreement",
        "parameters": {
          "title_number": "$response.body#/title_number"
        }
      },
      "DraftSalesAgreement": {
        "operationId": "add_sales_agreement_to_title",
        "parameters": {
          "title_number": "$response.body#/title_number"
        }
      },
      "SignSalesAgreement": {
        "operationId": "update_sales_agreement",
        "parameters": {
          "title_number": "$response.body#/title_number"
        }
      },
      "AddRestriction": {
        "operationId": "add_restriction_to_title",
        "parameters": {
          "title_number": "$response.body#/title_number"
        }
      },
      "RemoveRestrictions": {
        "operationId": "remove_restrictions_from_title",
        "parameters": {
          "title_number": "$response.body#/title_number"
        }
      },
...
    }
  }
}